+++ common/head_tag.html
@@ -10,5 +10,5 @@
     {{mount-widget widget="maker-badges-loader-widget"}}
 </script>
 
-<script src="/common/discourse-theme-component.js" type="text/discourse-plugin" version="0.8">
+<script src="/javascript/discourse-theme-component.js" type="text/discourse-plugin" version="0.8">
 </script>
